// src/interfaces/Course.interface.ts
export interface Course {
    id: number;
    name: string;
    link: string;
    category: string[];
}

/*
* 		"Робототехника": 1
		"Создание игр": 3
		"Программирование": 4
		"Мультимедиа": 3
		"Soft skills": 4
		"Шахматы": 1
		"Блогинг": 2
		"Web-разработка": 1
		"Прикладное программирование": 1
		"Графическое моделирование": 1
		"Game Development": 1
		"Разработчик мобильных приложений": 1
		"Разработчик игр на Unity": 1
		"Разработка ПО": 1
		"Графический Дизайн": 1
		"Тестировщик ПО": 1
		"Интернет Маркетинг": 1
		"Java-разработчик": 1
		"Python–разработчик": 1
		"IOS-разработчик": 1
* */